check shrinked image exists and then remove the backup file.

+++ img-shrink.sh
@@ -10,4 +10,7 @@ fi
 
 mv $1 $1.backup
 sudo qemu-img convert -O qcow2 -c $1.backup $1
-rm -f $1.backup
+
+if [ -f $1 ]; then
+  rm -f $1.backup
+fi
